Listas enlazadas de diferentes de los diferentes tipos funcinoando en java

+++ java/Java/src/main/java/com/wilsoncys/compi1/java/model/simbolo/Tipo.java
@@ -32,7 +32,7 @@ public class Tipo {
             case ENTERO:
                 return "int";
             case DECIMAL:
-                return "double";    //podria ser real
+                return "float";    //podria ser real
                 
             case CARACTER:
                 return "char";
